Add functions to plot CS radial hoop and radial stress profiles
